+++ CHANGELOG.md
@@ -1,7 +1,9 @@
 # Changelog
 
 ## Next
-- Add left arrow key shortcut for going to the parent playlist folder
+- Sidebar selection improvements:
+  - Add left arrow key shortcut for going to the parent playlist folder
+  - Fix up arrow selecting siblings only
 
 ## 0.16.12 - 2023 Jun 14
 - Improve filtering by perofrming unicode normalization