+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 - Sidebar selection improvements:
   - Add left arrow key shortcut for going to the parent playlist folder
   - Fix up arrow selecting siblings only
+  - Add alt+up/down shortcut
 
 ## 0.16.12 - 2023 Jun 14
 - Improve filtering by perofrming unicode normalization